#3606 added setting grid column (#3615)
