+++ discovery-frontend/src/app/common/component/chart/type/grid-chart.component.ts
@@ -319,11 +319,17 @@ export class GridChartComponent extends BaseChart implements OnInit, OnDestroy, 
     let aggregations: any;
     if( this.pivot && this.pivot.aggregations && 0 < this.pivot.aggregations.length ) {
       aggregations = this.pivot.aggregations.map((pivot) => {
+        const aggrInfo = {
+          name : _.isUndefined(pivot.alias) ? pivot.name : pivot.alias,
+          digits: 2
+        };
         if( pivot.field && pivot.field.logicalType ) {
-          return {name : _.isUndefined(pivot.alias) ? pivot.name : pivot.alias, digits: 2, type : pivot.field.logicalType };
-        } else {
-          return {name : _.isUndefined(pivot.alias) ? pivot.name : pivot.alias, digits: 2 };
+          aggrInfo['type'] = pivot.field.logicalType;
         }
+        if( pivot.fieldFormat ) {
+          aggrInfo['fieldFormat'] = pivot.fieldFormat;
+        }
+        return aggrInfo;
       });
     } else {
       aggregations = this.fieldInfo.aggs.map((name) => {
@@ -383,6 +389,16 @@ export class GridChartComponent extends BaseChart implements OnInit, OnDestroy, 
         if( pivot.field && pivot.field.logicalType ) {
           ( originAggregations[0].type ) || ( originAggregations[0].type = {} );
           originAggregations[0].type[ pivot.name ] = pivot.field.logicalType;
+
+          if( pivot.fieldFormat ) {
+            ( originAggregations[0]['fieldFormat'] ) || ( originAggregations[0]['fieldFormat'] = [] );
+            if( -1 === originAggregations[0]['fieldFormat'].findIndex( item => pivot.field.name === item['aggrColumn'] ) ) {
+              const fieldFormat = JSON.parse(JSON.stringify(pivot.fieldFormat));
+              fieldFormat['aggrColumn'] = pivot.field.name;
+              originAggregations[0]['fieldFormat'].push( fieldFormat );
+            }
+          }
+
         }
       });
       aggregations = originAggregations;
